Fix compilation error pointer comparison
The if statement here tested the pointer as a boolean (eg not null or
void) and also compared > 0. The latter part will not compile because it
is an arithmetic comparison of pointer to int and makes no sense. One
could change the second comparison to perimeter1[NumFire] != nullptr on
modern C++ (> C++11), but I can't see how that second portion would ever
be called as it is protected by the first test and a null pointer is
converted implicitly to FALSE.