Move src and targets to own directories
Cleaner for version control. I also removed Far_WN.cpp as I did in my
original fork of the 2016 posted farsite code. This code is not used in
the executable here and the firelab github repo README indicates that it
was based on a now obsolete version of WindNinja API. I quote from that
README: "Instead of trying to get that set of code to work, we took the
ideas from Assembla farsite to put in functionality for gridded wind
files so you can just use output from WindNinja as inputs to Farsite."