Add assert statements to confirm unused code
- option to perturb parameters slightly in DeadFuelMoisture class is not
used by the cli code. These assert statements could be removed This
class is fine as is and is well-organized and clear. I can confirm the
option is not used.

- I removed resetVisPerimFile() method from Farsite5 class. It did
nothing it returned nothing but also had no side effects because code
had already been commented-out. So the random file name creation was a
waste.

- I added assert statements to attack code to confirm code is unused.