Remove rand3.cpp/Random class
Move to using std::random. Generator for now is Mersenne twister random
number engine, mt19937. This addresses gh issue #1

farsite cli does not really use a lot of random number generation.
Spotting seems to be the main use? But my understanding of the code is
still weak.

Some parts of the code still use std:rand(). But those parts are either
not reachable by the CLI executable or are for non monte carlo non
critical applications.