Remove more unused files
splinex.h contained definitions for several functions but none of these
were used in any of the code. One cpp file had commented out code that
used one interpolation function, `polint`. I assume this and the other
functions were same/similar to those of the same names found in
Numerical Recipes in C.

The splinex.h file was indirectly bringing in <stdlib.h> and therefore
the definition of NULL. I replaced the one use that affected with
nullptr (C++11), but there are many other uses of NULL in the code.

farsite.hpp was never included in anything so removed.