Working threaded implementation of console app
Working version with progress notification and each command line in
command file running in a separate process thread. This all works, but
note that there is no thread pooling done, so the user should consider
how many lines to include in a single command file.

I'm not sure this complexity adds much. A simpler version would be to
have a progress thread and a main thread and only allow one command
line. This would allow/force steering code in scripts to handle
dispatching multiple simulations.