Some more minor changes prepping progress class
Not yet really using progress class. One issue is that fuel conditioning
takes a while but does not update progress percentage. Would be nice if
it could. Main function seems to be `int Farsite5::Run_CondDLL()` in
Far_Cond.cpp. So this is a member function and could access the
progress.