Move theme.h and gridthem.cpp to unused
As far as I can tell, this are pure frontend/graphics code. The only
place in the entire codebase these are used is lines 9645 - 9667 of
Farsite5.cpp. I can't see that those lines actually accomplish anything.
They call CI:GetFuelConversion from FMC_CI.h but then do nothing with
that information. I think this was all left in on accident?