Allow conditioning code to advance progress counter
This did result in some loss of compartmentalization: the FE2 class now
needs to know about the FPC (farsite progress) class. I did it by
passing a pointer from the FPC class object that each farsite run
containts down to the CFMC class and fromt here down to the FE2 object.

I'm not crazy about this, but it works