Sort of working progress bars
Uses indicators library: https://github.com/p-ranav/indicators

But DynamicProgress is weird and takes a references to the individual
bars not a copy (so bars created on stack can't be in a loop. I solved
this by putting bars on heap with new and storing in a array for later
deletion but this is clunky.

Also, no easy way to turn bar into a spinner for writing outputs stage.